collate unknown function on wk/ant w shared funcs
