+++ src/rard/research/models/antiquarian.py
@@ -8,7 +8,7 @@ from simple_history.models import HistoricalRecords
 from rard.research.models.mixins import HistoryModelMixin, TextObjectFieldMixin
 from rard.utils.basemodel import BaseModel, DatedModel, LockableModel, OrderableModel
 from rard.utils.decorators import disable_for_loaddata
-from rard.utils.shared_functions import collate_unknown
+from rard.utils.shared_functions import combine_links
 from rard.utils.text_processors import make_plain_text
 
 
@@ -385,6 +385,19 @@ class Antiquarian(
             for book in work.book_set.all():
                 if book.introduction:
                     book.introduction.link_bibliography_mentions_in_content()
+
+
+@disable_for_loaddata
+def collate_unknown(instance):
+    """This makes sure there's only one unknown work per antiquarian and combines contents if otherwise"""
+    unknown_works = instance.works.filter(unknown=True)
+
+    if unknown_works.count() > 1:
+        designated_unknown = unknown_works.first()
+        other_unknown_works = unknown_works.exclude(pk=designated_unknown.pk)
+
+    combine_links(other_unknown_works, instance, designated_unknown)
+    other_unknown_works.delete()
 
 
 @disable_for_loaddata
