+++ src/rard/research/models/work.py
@@ -16,7 +16,7 @@ from rard.research.models.base import (
 from rard.research.models.mixins import HistoryModelMixin, TextObjectFieldMixin
 from rard.utils.basemodel import BaseModel, DatedModel, LockableModel, OrderableModel
 from rard.utils.decorators import disable_for_loaddata
-from rard.utils.shared_functions import collate_unknown
+from rard.utils.shared_functions import combine_links
 from rard.utils.text_processors import make_plain_text
 
 
@@ -292,6 +292,20 @@ class Book(
                 if link.order_in_book != count:
                     link.order_in_book = count
                     link.save()
+
+
+@disable_for_loaddata
+def collate_unknown(instance):
+    """This makes sure there's only one unknown book per work and combines contents if otherwise"""
+    unknown_books = instance.book_set.filter(unknown=True)
+
+    if unknown_books.count() > 1:
+        designated_unknown = unknown_books.first()
+        other_unknown_books = unknown_books.exclude(pk=designated_unknown.pk)
+
+    combine_links(other_unknown_books, instance, designated_unknown)
+
+    other_unknown_books.delete()
 
 
 @disable_for_loaddata
